Add files via upload
Added backend routing: 
get requests to:
/index
/getpoints
/leaderboard
/tasks

made links to all static entities of the project to the public folder, ALL CSS and JS (related to DOM) should be uploaded in that folder,  all images to be uploaded in public/img folder. These static entities to be accessed directly by writing /css/<css file> or /js/<js file> or /img/<img file> instead of the complete relative path. and href links to be given as /<routing path> directly for backend compatibility.